Fix delete command
This was error response when trying to use delete method:

    Client command was not well formatted

Looking at packet sniffer, noticed client was sending:

    delete Id(1)

Notice the extraneous Id. Fixed by adding a custom display and now it
properly sends which adheres to the beanstlakd protocol.

    delete 1

May need to fix the other commands that send job ID as well.